empty

+++ scas/src/scas/structure/commutative/impl/Quotient.scala
@@ -7,7 +7,7 @@ import scas.structure.commutative.Quotient.Element
 
 trait Quotient[T](using ring: UniqueFactorizationDomain[T]) extends Field[Element[T]] {
   def apply(n: T) = Element(n, ring.one)
-  def apply(n: Long) = this(ring(n))
+  def apply(n: BigInteger) = this(ring(n))
   def apply(n: T, d: T): Element[T] = this(Element(n, d))
   override def convert(x: Element[T]) = {
     val Element(n, d) = x
