+++ scas/src/scas/structure/commutative/impl/Residue.scala
@@ -1,8 +1,10 @@
 package scas.structure.commutative.impl
 
+import scas.base.BigInteger
+
 trait Residue[T](using ring: UniqueFactorizationDomain[T]) extends UniqueFactorizationDomain[T] {
   def apply(x: T): T
-  def apply(n: Long) = this(ring(n))
+  def apply(n: BigInteger) = this(ring(n))
   override def convert(x: T) = this(ring.convert(x))
   extension (x: T) {
     def signum = ring.signum(x)
